immutable.rs, present_target.rs done except for raii vkDevice design (next trying vulkano again...

+++ src/renderer/present_target.rs
@@ -1,6 +1,7 @@
 // todo commenting
 // todo unwraps
 
+use crate::immutable::Immutable;
 use ash::{
     extensions::khr::{Surface, Swapchain},
     vk,
@@ -8,74 +9,93 @@ use ash::{
 pub use ash::{Device, Instance};
 use winit::window::Window;
 
-fn calc_work_group_count(image_size: vk::Extent2D) -> [u32; 2] {
-    let mut group_count_x = image_size.width / 16;
-    if (image_size.width % 16) != 0 {
-        group_count_x = group_count_x + 1;
-    }
-    let mut group_count_y = image_size.height / 16;
-    if (image_size.height % 16) != 0 {
-        group_count_y = group_count_y + 1;
-    }
-    [group_count_x, group_count_y]
-}
-
 pub struct PresentTarget {
-    // todo weak device pointers?
-    // References
-    device: Device,
-    physical_device: vk::PhysicalDevice,
-    surface_loader: Surface,
-    surface: vk::SurfaceKHR,
+    // Immutable members describing vulkan environment
+    // todo weak device ptr?
+    device: Immutable<Device>,
+    physical_device: Immutable<vk::PhysicalDevice>,
+    surface_loader: Immutable<Surface>,
+    surface: Immutable<vk::SurfaceKHR>,
+    swapchain_loader: Immutable<Swapchain>, // contains instance functions for swapchain extension
 
-    // todo getters!
-    pub swapchain_loader: Swapchain, // (doesn't impl Default unfortunately)
-    pub swapchain: vk::SwapchainKHR,
-    pub swapchain_image_views: Vec<vk::ImageView>,
-
-    pub image_count: u32,
-    pub resolution: vk::Extent2D,
-    pub work_group_count: [u32; 2],
+    // Presentation objects and info
+    swapchain: vk::SwapchainKHR,
+    swapchain_image_views: Vec<vk::ImageView>,
+    format: vk::SurfaceFormatKHR,
+    resolution: vk::Extent2D,
+    image_count: u32,
 }
 
+// Getters
+impl PresentTarget {
+    pub fn swapchain_loader(&self) -> &Swapchain {
+        &self.swapchain_loader
+    }
+    pub fn format(&self) -> vk::SurfaceFormatKHR {
+        self.format
+    }
+    pub fn swapchain(&self) -> &vk::SwapchainKHR {
+        &self.swapchain
+    }
+    pub fn swapchain_image_views(&self) -> &Vec<vk::ImageView> {
+        &self.swapchain_image_views
+    }
+    pub fn image_count(&self) -> u32 {
+        self.image_count
+    }
+    pub fn resolution(&self) -> vk::Extent2D {
+        self.resolution
+    }
+}
+
+// Public functions
 impl PresentTarget {
     pub fn new(
         window: &Window,
-        instance: Instance,
+        instance: &Instance,
         device: Device,
-        physical_device: vk::PhysicalDevice,
         surface_loader: Surface,
+        physical_device: vk::PhysicalDevice,
         surface: vk::SurfaceKHR,
     ) -> Self {
         let mut present_target = PresentTarget {
-            device,
-            physical_device,
-            surface_loader,
-            surface,
-            swapchain_loader: Swapchain::new(&instance, &device),
-            /* set in create() */ image_count: u32::default(),
-            /* set in create() */ resolution: vk::Extent2D::default(),
-            /* set in create() */ work_group_count: [u32::default(); 2],
+            swapchain_loader: Immutable::new(Swapchain::new(instance, &device)),
+            device: Immutable::new(device),
+            physical_device: Immutable::new(physical_device),
+            surface_loader: Immutable::new(surface_loader),
+            surface: Immutable::new(surface),
             /* set in create() */ swapchain: vk::SwapchainKHR::default(),
             /* set in create() */ swapchain_image_views: Vec::default(),
+            /* set in create() */ format: vk::SurfaceFormatKHR::default(),
+            /* set in create() */ resolution: vk::Extent2D::default(),
+            /* set in create() */ image_count: u32::default(),
         };
-        unsafe {
-            present_target.create(window);
-        }
+        present_target.create(window);
         present_target
     }
 
+    /// Recreate the swapchain due to e.g. window resize
     pub fn recreate(&mut self, window: &Window) {
         self.cleanup();
         self.create(window);
     }
 
+    /// Destroy vulkan objects owned by this struct and reinitialize data fields
+    pub fn cleanup(&mut self) {
+        self.destroy_vk_objs();
+        self.set_defaults();
+    }
+}
+
+// Private functions
+impl PresentTarget {
+    /// Create swapchain/images and set render target data
     fn create(&mut self, window: &Window) {
         unsafe {
             // present mode
             let present_mode = self
                 .surface_loader
-                .get_physical_device_surface_present_modes(self.physical_device, self.surface)
+                .get_physical_device_surface_present_modes(*self.physical_device, *self.surface)
                 .unwrap()
                 .iter()
                 .cloned()
@@ -83,15 +103,15 @@ impl PresentTarget {
                 .unwrap_or(vk::PresentModeKHR::FIFO);
 
             // choose format todo preference
-            let surface_format = self
+            self.format = self
                 .surface_loader
-                .get_physical_device_surface_formats(self.physical_device, self.surface)
+                .get_physical_device_surface_formats(*self.physical_device, *self.surface)
                 .unwrap()[0];
 
             // use surface capabilities data
             let surface_capabilities = self
                 .surface_loader
-                .get_physical_device_surface_capabilities(self.physical_device, self.surface)
+                .get_physical_device_surface_capabilities(*self.physical_device, *self.surface)
                 .unwrap();
 
             // resolution
@@ -122,10 +142,10 @@ impl PresentTarget {
             };
 
             let swapchain_ci = vk::SwapchainCreateInfoKHR::builder()
-                .surface(self.surface)
+                .surface(*self.surface)
                 .min_image_count(self.image_count)
-                .image_color_space(surface_format.color_space)
-                .image_format(surface_format.format)
+                .image_color_space(self.format.color_space)
+                .image_format(self.format.format)
                 .image_extent(self.resolution)
                 .image_usage(vk::ImageUsageFlags::COLOR_ATTACHMENT)
                 .image_sharing_mode(vk::SharingMode::EXCLUSIVE)
@@ -149,7 +169,7 @@ impl PresentTarget {
                 .map(|&image| {
                     let create_view_info = vk::ImageViewCreateInfo::builder()
                         .view_type(vk::ImageViewType::TYPE_2D)
-                        .format(surface_format.format)
+                        .format(self.format.format)
                         .components(vk::ComponentMapping {
                             r: vk::ComponentSwizzle::R,
                             g: vk::ComponentSwizzle::G,
@@ -172,23 +192,32 @@ impl PresentTarget {
         }
     }
 
-    fn cleanup(&mut self) {
-        //todo device checkz
+    /// Destroy vulkan objects owned by this struct
+    fn destroy_vk_objs(&mut self) {
+        todo!("device check");
         unsafe {
             for &image_view in self.swapchain_image_views.iter() {
                 self.device.destroy_image_view(image_view, None);
             }
-            self.swapchain_loader
-                .destroy_swapchain(self.swapchain, None);
+            if self.swapchain != vk::SwapchainKHR::default() {
+                self.swapchain_loader
+                    .destroy_swapchain(self.swapchain, None);
+            }
         }
+    }
+
+    /// Set default values for non-immutable fields
+    fn set_defaults(&mut self) {
+        self.swapchain = vk::SwapchainKHR::default();
+        self.swapchain_image_views = Vec::default();
+        self.format = vk::SurfaceFormatKHR::default();
         self.resolution = vk::Extent2D::default();
-        self.work_group_count = [0, 0];
+        self.image_count = u32::default();
     }
 }
 
 impl Drop for PresentTarget {
     fn drop(&mut self) {
-        self.cleanup();
-        //todo
+        self.destroy_vk_objs();
     }
 }
